requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
asyncpg==0.29.0
python-multipart==0.0.6
pydantic==2.4.2
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
cryptography==42.0.5
python-jose
cryptography